Added image functionality to blog
Images must be added to "blog/images" and can be called in a post by
![caption]({{ site.images }}/filename.ext)

There is also a front img field, so an img can be totally associated
with a post. Simply write the filename as a field in the post header.

---
layout: post
title: Title
img: file.png
---

Some slight styling issues in pseudosmalltablet. Will fix tomorrow.
Lookin' good!