Taking advantage of _includes
Things look a little crazy, pulling things in from all over. It's
reminding me of the article page modules. But it works, and it looks
good.

Some conceptual issues (particularly about the left-side navigation).
Refer to personal notes repo for more info.

Site is ready to have content plugged in, at which point it will be
ready for release.