Rethought navigation
Left-side navigation was only helpful in the standard context (and
maybe not even then). It's been replaced with a more explicit static
navigation on top of the docs, with sticky prev/next buttons and "back
to top" links that jump to full navigation.

On the blog, the tagNav is replaced with just a list of important tags.